Moving test to new file; reverting unnecessary changes to createOne test file
